revert changes from int to long for variables related to byteIndex

+++ core/src/main/java/org/apache/spark/shuffle/sort/ShuffleInMemorySorter.java
@@ -177,8 +177,8 @@ final class ShuffleInMemorySorter {
     if (useRadixSort) {
       offset = RadixSort.sort(
         array, (long)pos,
-        (long)PackedRecordPointer.PARTITION_ID_START_BYTE_INDEX,
-        (long)PackedRecordPointer.PARTITION_ID_END_BYTE_INDEX, false, false);
+        PackedRecordPointer.PARTITION_ID_START_BYTE_INDEX,
+        PackedRecordPointer.PARTITION_ID_END_BYTE_INDEX, false, false);
     } else {
       MemoryBlock unused = new MemoryBlock(
         array.getBaseObject(),
